refactor: unify TaskStatus into packages/types, remove client local redefinition
